requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
faiss-cpu==1.8.0.post1
numpy>=1.26
sentence-transformers==3.0.1
torch==2.3.1
llama-cpp-python==0.2.90